+++ src/TIVarTypes.h
@@ -24,12 +24,12 @@ namespace tivars
 
         static bool isValidName(const std::string& name);
 
-        static bool isValidID(int id);
+        static bool isValidID(uint8_t id);
 
         static std::vector<std::string> getExtensionsFromName(const std::string& name);
-        static std::vector<std::string> getExtensionsFromTypeID(int id);
+        static std::vector<std::string> getExtensionsFromTypeID(uint8_t id);
         static int getIDFromName(const std::string& name);
-        static std::string getNameFromID(int id);
+        static std::string getNameFromID(uint8_t id);
 
     private:
         static void insertType(const std::string& name, int id, const std::vector<std::string>& exts, const handler_pair_t& handlers = make_handler_pair(DummyHandler));
